requirements.txt
gradio
pytesseract
Pillow
python-docx
google-generativeai
pytz